Adicionando Telas CRUD

+++ CRUD/Models/EnderecoModel.cs
@@ -5,7 +5,7 @@
         protected int Id { get; set; }
         public string Logadouro { get; set; }
         public string Numero { get; set; }
-        public int Cep { get; set; }
+        public long Cep { get; set; }
         public string Bairro { get; set; }
         public string Cidade { get; set; }
         public string Estado { get; set; }
